feat(descriptionRenderer): Substitui a função parseDescription pela nova função parseText para melhorar a análise de descrições e corrige a ortografia de descripition para description em vários componentes
